chore(rig-rmcp): move inline test modules to sibling files
